skeleton/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
contourpy==1.0.7
cycler==0.11.0
dataclasses-json==0.5.7
filelock==3.12.0
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2023.5.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.15.1
idna==3.4
importlib-resources==5.12.0
kiwisolver==1.4.4
langchain==0.0.177
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
numexpr==2.8.4
numpy==1.24.3
openai==0.27.4
openapi-schema-pydantic==1.2.4
packaging==23.1
Pillow==9.5.0
pydantic==1.10.7
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==20.2
PyYAML==6.0
regex==2023.5.5
requests==2.31.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==1.4.48
telegram==0.0.1
tenacity==8.2.2
tiktoken==0.4.0
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==1.26.6
wikipedia==1.4.0
yarl==1.9.2
zipp==3.15.0
